fix: cant retry item in uploading batch
